perf: Avoid closing the socket prematurely
It seems to be possible to close the connection before the
socket's buffer is flushed. It's also very much possible to close
it before we have received repsonses to all of the commands sent.
We should actually wait for all the responses to really know how
long it all takes.