Always keep track of the SelvaObjectKey container
Unfortunately the commit 99be372ccb03a2099a3dc10aa300efa9cab760b4
broke more than it fixed. It's really hard to reliably track the
container object of the key and passing a wrong `obj` to the
`remove_key()` function may lead to catastrophic loss of data.

For now it will be much safer to always keep the pointer around in
the `SelvaObjectKey` struct. This will add 8 bytes to the
mentioned struct and 32 bytes to the `SelvaObjectKey` struct type
due to it containing 4 embedded keys. In that case it's
redundant to have that pointer but we only have one key type for
now.