Allow alternative server message handlers
Sometimes we want to send responses to other destinations than a
conn socket. E.g. previously we hacked our way through to discard
responses to replication messages.

Implementing a vtable for the message handling allows us to change
how to handle the responses. This allows us to implement not only a
discarding handler but also a handler that writes everything into a
single buffer for further processing.