Rework how hierarchy change events are sent
Instead of relying on special functions and magically appearing flags,
we should use the same field change logic for all fields. A hierarchy
change event should be sent if a field that is traversed in the marker
is changed. We keep the `SELVA_SUBSCRIPTION_FLAG_CH_HIERARCHY` flag to
mark that the marker is indeed a `change_marker` and to signal the
client that the change event was a hierarchy change.

Closes #181