Optimize sync load/save for replication
When doing a sync load or save we already have the filename and hash in
memory. We just need to pass those to the called of `selva_io_end()` to
be able to pass the information to the replication system.

This will remove an extra `fopen()`, some reading and some mallocs.