Avoid pointing last_good link to itself
If the user attempts to load last_good by sending the command
`load dump.sdb` this call shouldn't overwrite the symlink to
itself. Moreover, the actual dump filename should be determined
for better replication operation.

In addition, it might be a good idea to disallow '/' chars in the
filename. This also makes it even harder to traverse the file system and
write somewhere where the user shouldn't have access to.

Fixes #308