Fix indexing for has filters containing arrays (#326)
When we implemented the `has` filter by *has* we meant "one of these
items exists in the field X", but when we implemented indexing we
interpreted *has* as "is the field X a subset of
this set Y".
    
I think if we were mathematically exact, then has would check if
the set Y is a member of the set in the field X, but obviously that
has no meaning in selva. As the situation is what it is and the
meaning of *has* is now fixed, we just need to fix the indexing
hint.
    
Unfortunately we don't currently have a set-like relation feature
for the `has`. A good enough and pretty performant alternative is
to use the possibly (◇, Q) and drop/discard (U) operators together
to short-circuit when a match is found. This will result hints
in the following format when an array is used with `has`:
    
```
'"th1" "subthings" a Q U "th5" "subthings" a'
```

Fixes #324