Move the read looping to the event loop
We gained a lot of speed with the latest perf improvements looping
on `read` instead of going back to poll after receiving one frame
from each client. However, this broke the fairness of the server as it's
now blocking all other clients until the first one is ready.

By moving the looping to the event loop and making it conditional it's
possible to RR between those event handlers that still want to receive
more data. This can be thought as a kind of cooperative multi-tasking.