Keep track of the greatest id in a refs field
We have defined that ids start from 0 -> INT32_MAX. We can pretty much
assume that typically we are inserting ids in the creation order to
references fields, i.e. 1, 2, 3,.. This means that if we know the
largest id ever inserted we know wether there can be a "collision".
Therefore, we know that the lookup is only necessary if the new
insertion has a nodeId smaller than the largest currently in the "set".
I believe that in majority of the cases where we use extremely large
references fields we are mostly inserting almost every new node. In the
use cases where nodeIds are more randomly distributed we probably tend
to have shorter references fields anyway, that's something like
"friends", "contributors", or "editors" lists, that are trivial to just
loop O(n).

Ref BASED-282