Most global variables must be in TLS
Now that we use worker threads the global variables that are modified in
multiple threads must be stored in TLS. An exception is most of `db.c`
that is designed to work well and without locks for multi-threading. On
the other hand, particularly anything using global instances of
`libdeflate` compressors and decompressors must use thread-local
instances.

Ref FND-499