Option to enable natural order for mempools
Add `MEMPOOL_GROWING_FREE_LIST` option to make mempool sort free
chunk in init from the smallest to the greatest address.
It looks like at least in the naive tests the original behaviour is
faster, probably due to the fact that the last chunks are cached.
Hence, the option is unset for now.