Don't crash if a ref was already created
The check is inherited from the previous version of the refs code and
it should be now actually ok to fail on writing to the index and just
skip the 2-way write. An error in the 2-way write is still fatal.

```
fields.c:1228:selva_fields_references_insert_tail_wupsert_nonempty_src_field:
Can't happen, right
```