Simplify fields writing
`fields_set()` and `fields_set2()` are kinda dumb because the functions
usually hide the actual type and the caller is likely doing switch..case
anyway, thus the functions will just do the same switch..case for the
type again. Moreover, half of the types can't be even set with these
functions anyway.

Ref FDN-1291